top bar re-org
